wkg on sending 32 bit data back to RPi2B debug in even_odd_de_inter      71                         if (bp == 0x3ffffffd)      72                         zip_halt(); issue compareing data created zipcpu and RPi2B

+++ zipcpu/uart_test_code/disp.m
@@ -1,8 +1,8 @@
 %to load load "disp.m"
 %to run disp
-fid = fopen('r_dwt_zip.bin','r'); im = fread(fid, [256,inf], 'int16'); fclose(fid);
-fid = fopen('g_dwt_zip.bin','r'); im1 = fread(fid, [256,inf], 'int16'); fclose(fid);
-fid = fopen('b_dwt_zip.bin','r'); im2 = fread(fid, [256,inf], 'int16'); fclose(fid);
+fid = fopen('r_dwt_zip.bin','r'); im = fread(fid, [256,inf], 'int32'); fclose(fid);
+fid = fopen('g_dwt_zip.bin','r'); im1 = fread(fid, [256,inf], 'int32'); fclose(fid);
+fid = fopen('b_dwt_zip.bin','r'); im2 = fread(fid, [256,inf], 'int32'); fclose(fid);
 figure
 imagesc(im)
 title("Red subband 3 levels dwt ZipCpu")
